+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
 ---
+## TBD
+- Implemented `Policy` trait to adjust Glyph behavior.
+
+---
 ## 1.0.0: 2022-11-19
 - Released.